IDEN-486 Move all test vector utils to one file
